% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bacteria_dist}
\alias{bacteria_dist}
\title{Bacteria phylogenetic distances}
\format{
\subsection{\code{bacteria_dist }}{

A matrix with ...
}
}
\source{
ape package
}
\usage{
bacteria_dist
}
\description{
Square matrix of distances
}
\keyword{datasets}
