% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fleas_dist}
\alias{fleas_dist}
\title{Flea phylogenetic distances}
\format{
\subsection{\code{fleas_dist }}{

A matrix with 202 rows and columns
}
}
\source{
ape package
}
\usage{
fleas_dist
}
\description{
Square matrix of distances
}
\keyword{datasets}
