% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gopher.D}
\alias{gopher.D}
\title{Gopher phylogenetic distances}
\format{
\subsection{\code{gopher.D}}{

A matrix with 15 rows and columns
}
}
\source{
ape package
}
\usage{
gopher.D
}
\description{
Square matrix of distances
}
\keyword{datasets}
