% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gopher.lice.links}
\alias{gopher.lice.links}
\title{Gopher and lice association matrix}
\format{
\subsection{\code{gopher.lice.links}}{

A matrix with 15 rows and 17 columns
}
}
\source{
ape package
}
\usage{
gopher.lice.links
}
\description{
A matrix of associations between gopher and lice. 1 = associated, 0 = no association
}
\keyword{datasets}
