% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lice.D}
\alias{lice.D}
\title{Lice phylogenetic distances}
\format{
\subsection{\code{lice.D}}{

A matrix with 17 rows and columns
}
}
\source{
ape package
}
\usage{
lice.D
}
\description{
Square matrix of distances
}
\keyword{datasets}
