% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mammals_dist}
\alias{mammals_dist}
\title{Mammal phylogenetic distances}
\format{
\subsection{\code{mammals_dist }}{

A matrix with 130 rows and columns
}
}
\source{
ape package
}
\usage{
mammals_dist
}
\description{
Square matrix of distances
}
\keyword{datasets}
