% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mammals_fleas_links}
\alias{mammals_fleas_links}
\title{Mammal and flea association matrix}
\format{
\subsection{\code{mammals_fleas_links }}{

A matrix with 130 rows and 202 columns
}
}
\source{
ape package
}
\usage{
mammals_fleas_links
}
\description{
A matrix of associations between mammals and fleas. 1 = associated, 0 = no association
}
\keyword{datasets}
