\name{parafit-package}
\alias{parafit-package}
\docType{package}
\title{ Test of host-parasite coevolution }
\description{
Function \code{\link{parafit}} tests the hypothesis of coevolution between a clade of hosts and a clade of parasites. The null hypothesis (H0) of the global test is that the evolution of the two groups, as revealed by the two phylogenetic trees and the set of host-parasite association links, has been independent. Tests of individual host-parasite links are also available as an option.
}
\details{
\tabular{ll}{
Package: \tab parafit\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2009-05-14\cr
License: \tab Copyleft\cr
LazyLoad: \tab yes\cr
}
Function \code{\link{parafit}} tests the hypothesis of coevolution between a clade of hosts and a clade of parasites.

Function \code{\link{pcoa}} computes principal coordinate decomposition (classical scaling) of a distance matrix D and outputs the eigenvectors corresponding to all eigenvalues, positive and negative. It implements two correction methods for negative eigenvalues.
}

\author{ Pierre Legendre }

\references{
Legendre, P., Y. Desdevises and E. Bazin. 2002. A statistical test for host-parasite coevolution. Systematic Biology 51(2): 217-234.
}

\keyword{ package }
