% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{permute_cube_indices}
\alias{permute_cube_indices}
\title{Create a permutation cube}
\usage{
permute_cube_indices(input_matrix, permutations, seed)
}
\arguments{
\item{input_matrix}{matrix for dimensions}

\item{permutations}{number of permutations}

\item{seed}{random seed value}
}
\value{
Array with matrix permutations
}
\description{
Permute the values of each row independently, in multiple slices (permutations).
First permutation is the original indices.
}
