% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_parafit_ape.R
\name{pf_parafit_ape}
\alias{pf_parafit_ape}
\title{Parafit}
\usage{
pf_parafit_ape(
  host.D,
  para.D,
  HP,
  nperm = 999,
  test.links = FALSE,
  seed = NULL,
  correction = "none",
  silent = FALSE,
  parallel = F,
  cores = 1
)
}
\arguments{
\item{host.D}{transposed host principal coordinates}

\item{para.D}{parasite principal coordinates}

\item{HP}{association matrix}

\item{nperm}{number of permutations}

\item{test.links}{=FALSE,}

\item{seed}{=NULL,}

\item{correction}{="none",}

\item{silent}{=FALSE,}

\item{parallel}{= F,}

\item{cores}{= 1}
}
\value{
list of results, with elements global and links (the latter if test_links = TRUE)
}
\description{
Run the parafit algorithm to obtain the global test statistics and obtain a test
of significance using permutation. Optionally, obtain tests of individual associations.
}
