% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pf_parafit_cpp}
\alias{pf_parafit_cpp}
\title{Run the parafit algorithm}
\usage{
pf_parafit_cpp(assoA, paraB, hostC, permutations)
}
\arguments{
\item{assoA}{association matrix}

\item{paraB}{parasite principal coordinates}

\item{hostC}{transposed host principal coordinates}

\item{permutations}{number of permutations}
}
\value{
list of results, with elements global and links (the latter if test_links = TRUE).
}
\description{
Run the parafit algorithm to obtain the global test statistics and obtain a test
of significance using permutation. Optionally, obtain tests of individual associations.
}
