% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_parafit_r.R
\name{pf_parafit_r}
\alias{pf_parafit_r}
\title{Parafit}
\usage{
pf_parafit_r(assoA, paraB, hostC, permutations, seed)
}
\arguments{
\item{assoA}{association matrix}

\item{paraB}{parasite principal coordinates}

\item{hostC}{transposed host principal coordinates}

\item{permutations}{number of permutations}

\item{seed}{random seed value}
}
\value{
vector of results, with elements global and links (the latter if test_links = TRUE)
}
\description{
Run the parafit algorithm to obtain the global test statistics and obtain a test
of significance using permutation. Optionally, obtain tests of individual associations.
}
