% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pf_pcoa_cpp}
\alias{pf_pcoa_cpp}
\title{Obtain Principal Coordinate from a distance matrix}
\usage{
pf_pcoa_cpp(x)
}
\arguments{
\item{x}{a distance matrix}
}
\value{
matrix of principal coordinates
}
\description{
Obtain principal coordinates from a distance matrix (triangle or square)
}
