% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_parafit.R
\name{print.pf_parafit}
\alias{print.pf_parafit}
\title{Print method for pf_parafit class}
\usage{
\method{print}{pf_parafit}(x)
}
\arguments{
\item{x}{pf_parafit object}
}
\description{
Print method for pf_parafit class
}
