% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sponge_bacteria_links}
\alias{sponge_bacteria_links}
\title{Sponge and bacteria association matrix}
\format{
\subsection{\code{sponge_bacteria_links }}{

A matrix with ...
}
}
\usage{
sponge_bacteria_links
}
\description{
A matrix of associations between sponge and bacteria. 1 = associated, 0 = no association
}
\keyword{datasets}
