% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sponge_dist}
\alias{sponge_dist}
\title{Spnge phylogenetic distances}
\format{
\subsection{\code{sponge_dist }}{

A matrix with ...
}
}
\usage{
sponge_dist
}
\description{
Square matrix of distances
}
\keyword{datasets}
