% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{wrangle_pf_pcoa_cpp}
\alias{wrangle_pf_pcoa_cpp}
\title{Wrangle PCOA}
\usage{
wrangle_pf_pcoa_cpp(x)
}
\arguments{
\item{x}{POCA from ...}
}
\value{
list
}
\description{
Add additional objects to PCOA
}
