% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_pcoa.R
\name{pf_pcoa}
\alias{pf_pcoa}
\title{Obtain Principal Coordinate from a distance matrix}
\usage{
pf_pcoa(x, correction = "none", use_cpp = FALSE)
}
\arguments{
\item{x}{a distance matrix}

\item{correction}{description}

\item{use_cpp}{use ape pacakge or CPP}
}
\value{
matrix of principal coordinates
}
\description{
Obtain principal coordinates from a distance matrix (triangle or square)
}
